unify tests for blaslt, remove hipblas-common dependency but add it in as an include for nvidia CI. also dedup code
